module 3. task 5. fix

+++ src/main/java/lang/print/gaps/task5/NumberDivider.java
@@ -2,7 +2,7 @@ package lang.print.gaps.task5;
 
 public class NumberDivider {
     void divide(int toBeDivided) {
-        int res = toBeDivided / 5;
+        double res = toBeDivided / 5;
         System.out.println(res);
     }
 }
